+++ src/civilization/general/general_systems.rs
@@ -10,7 +10,6 @@ use bevy::prelude::{default, info, Commands, Entity, EventReader, EventWriter, N
 use bevy::sprite::SpriteBundle;
 use bevy_console::PrintConsoleLine;
 use rand::seq::IteratorRandom;
-use crate::civilization::trade_cards::trade_card_components::PlayerTradeCards;
 
 pub fn start_game(
     player_query: Query<(Entity, &Name, &Faction), With<Player>>,
